requirements.txt
requests>=2.31.0
python-dotenv~=0.21.0
rich~=12.6.0
selenium==4.15.0
supabase==1.1.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
pandas>=1.5.0
numpy>=1.24.0
beautifulsoup4>=4.11.0
lxml>=4.9.0